requirements.txt
Django==2.2.12
psycopg2==2.7.4
django-redis==4.10.0
wagtail==2.8.1
Pillow==6.2.2
embedly==0.5.0
requests==2.20.0
raven==6.9.0
wagtail-django-recaptcha==1.0
django-basic-auth-ip-whitelist==0.2.1
boto3==1.7.82
django-storages==1.6.6
-e git://github.com/wagtail/wagtail-review.git@8ec5ec188ee75a075bd83c3a7804aeac9abf71be#egg=wagtail-review
django-phonenumber-field==2.1.0
phonenumbers==8.10.3
graphene_django==2.2.0
django-cors-headers==3.0.2

# Development requirements
django-debug-toolbar==1.2.2

# Production dependencies
dj-database-url==0.5.0
whitenoise==5.0.1
ConcurrentLogHandler==0.9.1

# Wagtail-markdown
wagtail-markdown==0.6
Pygments==2.5.2  # for code highlighting in wagtail-markdown
wagtailfontawesome==1.2.1
scout-apm==1.1.7

scour==0.37
wagtail-headless-preview==0.1.4
wagtailgatsby
